Add support for first-time deployments
This changeset adds support to the deploy script for first-time deployments.  Previously, we were assuming an app was already deployed and performing the zero-downtime-push command of the autopilot plugin.  However, this will fail in the event that the app does not already exist, so this change defaults to performing a regular push command in the event the app is not already deployed.